+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ## [Unreleased]
 ### Added
 - Color settings for OBJ 3D preview colors for faces, lines and points.
+- Settings for the 3D preview:
+  - Default up vector
+  - Line width
+  - Point size
 
 ### Fixed
 - Separate color settings attribute key for OBJ and MTL constants